release 0.11.1
* message: add missing message property encoding skipped by c library

The Azure IoT Hub expects message properties to be url encoded. However, the C
azure-iot-sdk does not perform such an encoding (see also
https://github.com/Azure/azure-iot-sdk-python/issues/227 for a discussion of
this issue). The effect of this is, that setting e.g. the message content type
to "application/json" as suggested by the docstrings, leads to the IoT hub not
accepting them.

This commit takes care to urlencode message property keys and values.

* twin: fix lint warnings

* client: fix formatting

* release 0.11.1